migrations issue

+++ database/migrations/2020_05_05_090500_create_recordings_table.php
@@ -22,7 +22,7 @@ class CreateRecordingsTable extends Migration
             $table->foreign('user_id')->references('id')->on('users');
             $table->bigInteger('style_id')->unsigned()->nullable();
             $table->foreign('style_id')->references('id')->on('styles');
-            $table->bigInteger('title_id')->unsigned()->nullable();
+            $table->integer('title_id')->unsigned()->nullable();
             $table->foreign('title_id')->references('id')->on('titles');
             $table->timestamps();
         });
